refactor(board): rename import directory to obf
Rename src/features/board/import/ to src/features/board/obf/ to better reflect the directory's responsibility (Open Board Format mapping and import). Update internal import paths, the public barrel, and architecture docs accordingly. No behavior change.
