+++ src/features/board/index.ts
@@ -1,11 +1,11 @@
 export { BoardView } from "./BoardView";
-export type { BoardSetRecord } from "./storage/boards-db";
-export { useBoard } from "./useBoard";
-export { useBoardSets } from "./storage/useBoardSets";
-export { useImportBoardFiles } from "./import/useImportBoardFiles";
+export { useImportBoardFiles } from "./obf/useImportBoardFiles";
 export {
   fetchBoardSets,
   importBoardFromUrl,
   removeBoardSet,
 } from "./storage/board-sets-store";
+export type { BoardSetRecord } from "./storage/boards-db";
+export { useBoardSets } from "./storage/useBoardSets";
 export type { BoardRouteParams } from "./types";
+export { useBoard } from "./useBoard";
