+++ src/features/board/obf/obf-mapper.ts
